+++ lang/cem/cemcom/cstoper.c
@@ -230,8 +230,8 @@ init_cst()
 	}
 	mach_long_size = i;
 	mach_long_sign = 1L << (mach_long_size * 8 - 1);
-	if (long_size < mach_long_size)
+	if ((int)long_size < mach_long_size)
 		fatal("sizeof (long) insufficient on this machine");
-	max_int = full_mask[(int)int_size] & ~(1L << (int_size * 8 - 1));
+	max_int = full_mask[(int)int_size] & ~(1L << ((int)int_size * 8 - 1));
 	max_unsigned = full_mask[(int)int_size];
 }
